+++ internal/pkg/service/appsproxy/proxy/apphandler/authproxy/oauthproxy/config.go
@@ -1,4 +1,4 @@
-package oidcproxy
+package oauthproxy
 
 import (
 	"crypto/sha256"
