+++ internal/pkg/service/appsproxy/proxy/apphandler/authproxy/oauthproxy/doc.go
@@ -1,3 +1,3 @@
 // Package oidcproxy provides factory for OAuth2Proxy library.
 // The logic is inserted before the handler from  the "upstream" package.
-package oidcproxy
+package oauthproxy
