+++ internal/pkg/service/appsproxy/proxy/apphandler/authproxy/oauthproxy/handler.go
@@ -1,4 +1,4 @@
-package oidcproxy
+package oauthproxy
 
 import (
 	"fmt"
