+++ internal/pkg/service/appsproxy/proxy/apphandler/authproxy/oauthproxy/logging/writer.go
