+++ internal/pkg/service/appsproxy/proxy/apphandler/authproxy/oauthproxy/pagewriter.go
@@ -1,4 +1,4 @@
-package oidcproxy
+package oauthproxy
 
 import (
 	"net/http"
